==> Scripts/2. kyoto_Paris_Emission_country_year(Finalizado).ipynb <==
# Países que assinaram o Acordo de Paris

Este notebook tem o intuito de relacionar as bases 1 e 2 conforme descritos abaixo:
    
    1. "Emission_Country_Year" - Base principal do projeto, pois contén os dados quantitativos relacionados a emissão de gases poluentes por País e Ano, esta base registra dados de 1850 a 2019 de 236 países.

    2. "Paris_Agreement_Countries" - Base auxiliar do projeto, pois contém informações de países que assinaram o Acordo de Paris, esta base registra dados de 176 países. 

As relações das bases se darão pelas colunas Base 1, Coluna "country" e Base 2, coluna "Participant" que são comuns em ambas as bases, Foram necessárias algumas trtativas na Base 2 a fim de deixa-la o mais proximo possível da Base 1.

Contextualização: O Acordo de paris é um tratado internacional com compromissos mais rígidos para a redução da emissão dos gases que produzem o efeito estufa, que são a causa do atual aquecimento global, O acordo de Paris substitui o Tratado de Kyoto a Partir de 2020.

# Instalar os programas e bibliotecas que serão utilizados
import pandas as pd
import requests
from bs4 import BeautifulSoup

1. Base de Dados 2 - "Paris_Agreement_Countries"

#Requisição de dados no site para coleta de dados e cração da base 2. Kyoto_Agreement_Countries
reqparis = requests.get('https://treaties.un.org/pages/ViewDetails.aspx?src=TREATY&mtdsg_no=XXVII-7-d&chapter=27&clang=_en')
if reqparis.status_code == 200:
    print('Requisição bem sucedida!')
    contentparis = reqparis.content

#criação de lista com dados requisitados anteriormente
soupparis = BeautifulSoup(contentparis, 'html.parser')
#buscando a tag table na lista, já que os dados que vamos consumir está em uma tabela neste site.
tableparis = soupparis.find(name='table')

#transformando lista em string
table_str_paris = str(tableparis)
#lendo o 10° item da lista e o convertendo em dataframe, obs: foi testado varias posiçoes da lista e identificado que a tabela 10 é exatamente a tabela necessitada.
Paris_Agreement_Countries = pd.read_html(table_str_paris)[10:11][0]
#Drop na linha 0 do dataframe pois trata-se dos nomes da coluna, em caso de dúvidas olhar na visualização acima na linha de número 0
Paris_Agreement_Countries.drop(0, axis=0, inplace=True)
#exibição do dataframe
Paris_Agreement_Countries

#Drop NAN na coluna 2
Paris_Agreement_Countries.dropna(subset=[2], inplace=True)
#inserção dos nomes das colunas no dataframe
Paris_Agreement_Countries.columns=['Participant','Signature','Date']
#Criação da coluna 'Accession(a)' que vale True se a coluna 'Date' finaliza com ' a'
Paris_Agreement_Countries['Accession'] = Paris_Agreement_Countries.Date.str.endswith(" a")
#Criação da coluna 'Accession(a)' que vale True se a coluna 'Date' finaliza com ' a'
Paris_Agreement_Countries['Approval'] = Paris_Agreement_Countries.Date.str.endswith(" AA")
#Criação da coluna 'Accession(a)' que vale True se a coluna 'Date' finaliza com ' a'
Paris_Agreement_Countries['Acceptance'] = Paris_Agreement_Countries.Date.str.endswith(" A")
#Criação de coluna Ratification onde todas as demais valem False
Paris_Agreement_Countries['Ratification'] = (Paris_Agreement_Countries['Accession'] == False) & (Paris_Agreement_Countries['Approval'] == False) & (Paris_Agreement_Countries['Acceptance'] == False)
#exibição do dataframe
Paris_Agreement_Countries

A Coluna Signature se refere a data em que o país assinou o tratado de Kyoto, na coluna Date trata-se de diversas ações possívesi após a assinatura como:
   1. Approval(AA)
   2. Acceptance(A)
   3. Accession(a)
   4. Ratification
  
A informação mais importante nesse visão do projeto é a Ratificação e não apenas mais adesão pois apenas 2 países aderiram ao tratado conforme dataframe, os dados referentes a aceitação ou Aprovação serão descartados no momento, pois deseja-se descobrir quais países validaram o tratado.

#reset index
Paris_Agreement_Countries.reset_index(inplace=True)
#drop coluna index
Paris_Agreement_Countries.drop(['index'], axis=1, inplace=True)
#Drop coluna
Paris_Agreement_Countries.drop(['Accession'], axis=1, inplace=True)
#Drop coluna
Paris_Agreement_Countries.drop(['Approval'], axis=1, inplace=True)
#Drop coluna
Paris_Agreement_Countries.drop(['Acceptance'], axis=1, inplace=True)
# indexa sempre que a coluna 'Ratification' valer false, ou seja, quando não aderido.
indexNames = Paris_Agreement_Countries[(Paris_Agreement_Countries['Ratification'] == False)].index 
# Deletar linhas conforme indexação
Paris_Agreement_Countries.drop(indexNames, inplace=True)
#Reset index
Paris_Agreement_Countries.reset_index(inplace = True)
#Drop coluna index
Paris_Agreement_Countries.drop(['index'],axis=1, inplace = True)
#criação de novo dataframe
Paris_Agreement_Countries_in = Paris_Agreement_Countries
#exibição do dataframe
Paris_Agreement_Countries

#Drop coluna Signature
Paris_Agreement_Countries.drop(['Signature'], axis=1, inplace=True)
#SubstituiParis_Agreement_Countries a string ' a' por '' na coluna 'Date' quando a mesma finalizar com este valor 
Paris_Agreement_Countries.Date = Paris_Agreement_Countries.Date.str.replace(' a$', '')
#Substitui a string '[' por '' na coluna 'Date' quando a mesma contiver este valor 
Paris_Agreement_Countries.Date = Paris_Agreement_Countries.Date.str.replace('[', '')
#Substitui a string '[' por '' na coluna 'Date' quando a mesma contiver este valor 
Paris_Agreement_Countries.Date = Paris_Agreement_Countries.Date.str.replace(' ]', '')
#Substitui a string ' ' por '' na coluna 'Date' quando a mesma contiver este valor 
Paris_Agreement_Countries.Date = Paris_Agreement_Countries.Date.str.replace(' $', '')
#Drop todos os valores nulos do dataframe
Paris_Agreement_Countries.dropna(inplace=True)

#Converte a coluna 'Date' em data
Paris_Agreement_Countries['Date'] = pd.to_datetime(Paris_Agreement_Countries['Date'])
#Corrigido os nomes na coluna 'Participant' pois havia nomes com '()'
Paris_Agreement_Countries['Participant'] = Paris_Agreement_Countries['Participant'].str.split('(').apply(lambda x : x[0])
#Substitui a string ' ' por '' na coluna 'Participant' quando a mesma finalizar com este valor 
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace(' $', '')
#Reset index
Paris_Agreement_Countries.reset_index(inplace = True)
#Drop coluna index
Paris_Agreement_Countries.drop(['index'],axis=1, inplace = True)
#exibição do dataframe
Paris_Agreement_Countries

2. Base de Dados 1 - "Emission_Country_Year"

#Leitura de arquivo salvo na VM
Emission_Country_Year = pd.read_csv('/home/jupyter/Emission_Country_Year.csv')
#Drop coluna 'Unnamed: 0'
Emission_Country_Year.drop(['Unnamed: 0'],axis=1,inplace=True)
#criação de novo dataframe com os dados anteriores ao tratamento
Emission_Country_Year_in = Emission_Country_Year
#exibição do dataframe
Emission_Country_Year

Tratamento da base "1. Emission_Country_Year"

#Criação de dataframe com 2 colunas de Emission_Country_Year
Country_IsoCode = Emission_Country_Year[['iso_code', 'country', 'year']]
#Drop todos os valores nulos da coluna 'country' 
Country_IsoCode.dropna(subset=['country'], inplace=True)
#exibição do dataframe
Country_IsoCode

#Preencher sempre que Iso_code valer Nan
Country_IsoCode['iso_code'].fillna('Preencher',inplace=True)
#Agrupamento por 'iso_code' selecionando os dados da tabela 'country' index resetado para 'iso_code' e o index virou a coluna
Country_IsoCode = Country_IsoCode.groupby('country')['iso_code'].apply(lambda Country_IsoCode: Country_IsoCode.reset_index(drop=True)).unstack(0)
#Drop todos os valores nulos
Country_IsoCode.dropna(inplace = True)
#exibição do dataframe
Country_IsoCode

#dataframe transposto
Country_IsoCode = Country_IsoCode.transpose()
#reset de index
Country_IsoCode.reset_index(inplace = True)
#Substitui a string ' ' por '' na coluna 'country' quando a mesma finalizar com este valor 
Country_IsoCode.country = Country_IsoCode.country.str.replace(' $', '')
#Renomear nome das colunas
Country_IsoCode.columns=['Country','Iso_code']
#exibição do dataframe
Country_IsoCode

#verificando quantos valores da coluna 'Country são nulos'
Country_IsoCode['Country'].nunique()

3. Alteração na base "Base 2. Paris_Agreement_Countries", Conforme Base de Dados 1 - "Emission_Country_Year"

#Substituição dos nomes dos países a fim de deixar a base conforme a base principal do projeto 
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('Brunei Darussalam', 'Brunei')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('Cabo Verde', 'Cape Verde')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace("Côte d'Ivoire", "Cote d'Ivoire")
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('Democratic Republic of the Congo', 'Democratic Republic of Congo')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace("Lao People's Democratic Republic", 'Laos')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('Republic of Moldova', 'Moldova')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('St. Kitts and Nevis', 'Saint Kitts and Nevis')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('Syrian Arab Republic', 'Syria')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace("Democratic People's Republic of Korea", 'North Korea')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('Timor-Leste', 'Timor')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('United Republic of Tanzania', 'Tanzania')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('Czech Republic', 'Czechia') 
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('New Zealand 3', 'New Zealand') 
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('St. Lucia', 'Saint Lucia') 
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace('St. Vincent and the Grenadines', 'Saint Vincent and the Grenadines') 
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace("Republic of Korea", 'South Korea') 
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace("United Kingdom of Great Britain and Northern Ireland", 'United Kingdom')
Paris_Agreement_Countries.Participant = Paris_Agreement_Countries.Participant.str.replace("State of Palestine", 'Palestine')
#exibição do dataframe
Paris_Agreement_Countries

Drop de valores que não existiam na base de "1. Emission_Country_Year".

# Get indexes where name column doesn't have value john
indexNames = Paris_Agreement_Countries[(Paris_Agreement_Countries['Participant'] == 'San Marino') | (Paris_Agreement_Countries['Participant'] == 'Monaco') | (Paris_Agreement_Countries['Participant'] == 'European Union')].index 
# Delete these row indexes from dataFrame
Paris_Agreement_Countries.drop(indexNames, inplace=True)
#exibição do dataframe
Paris_Agreement_Countries

4. Análise de dados - Verificação se Base de Dados 1 - "Emission_Country_Year" e Base 2. "Paris_Agreement_Countries"estão correspondentes:

Foram Analisadas as colunas Paris_Agreement_Countries['Participant'] e Country_IsoCode['Country'] para verificação de correspondencias.
Foram criados 3 dataframes conforme abaixo:
    
    1. paris_naoIsocode: Países existentes na base de dados "2. paris_Agreement_Countries" porém não na base "1. Emission_Country_Year", foram encontrados 20 valores na primeira iteração, porém houve verificação na escrita destes países e houve alterações de nomes, conforme item 2.3, após estas alterações encontrou-se apenas 1 valor, como a base 2 é apenas uma base auxiliar este país foi excluído da base 2, conforme iteração 3.

#Dataframe criado com as linhas em que a coluna 'Participant' nao estava em Country_IsoCode['country'] nesse caso sera necessario substituir
#Paris_Agreement_Countries['Participant'] pela sua base e coluna a qual contem o nome dos países
Paris_naoIsocode = Paris_Agreement_Countries[Paris_Agreement_Countries['Participant'].isin(Country_IsoCode['Country']) == False]
#contando o numero de linhas nao casadas
Paris_naoIsocode.count()

Paris_naoIsocode

    2. Paris_Isocode: Países existentes em ambas as base, foram encontrados 98 valores na primeira iteração, porém após as alterações na base conforme item 2.3 esse valor subiu para 109 países, ou seja, 109 países da base geral eceitaram implementar ações para diminuição de emissão de gases poluentes.

#Dataframe criado com as linhas em que a coluna 'Participant' existe em, nesse caso temos o match e nao sera necessario se preocupar com estas linhas
Paris_Isocode = Paris_Agreement_Countries[Paris_Agreement_Countries['Participant'].isin(Country_IsoCode['Country'])]
#contando o numero de linhas casadas entre os dataframes
Paris_Isocode.count()

Paris_Isocode

    3. naoparis_Isocode: Países existentes na base de dados "1. Emission_Country_Year" porém não na base "2. Paris_Agreement_Countries", foram encontrados 138 valores na primeira iteração, porém após as alterações na base conforme item 2.3 esse valor diminuiu para 127 países, ou seja, 127 países da base "1. Emission_Country_Year" não implementaram ações para a diminuição na emissão de gases poluentes. 

#Dataframe criado com as linhas em que a coluna Country_IsoCode['country'] nao estava em Paris_Agreement_Countries['Participant'] nesse caso sera necessario substituir
#Paris_Agreement_Countries['Participant'] pela sua base e coluna a qual contem o nome dos países
naoParis_Isocode = Country_IsoCode[Country_IsoCode['Country'].isin(Paris_Agreement_Countries['Participant']) == False]
#contando o numero de linhas nao casadas
naoParis_Isocode.count()

naoParis_Isocode

5. Conclusão

Será incluído uma coluna no Dataframe 1 com os Valores True e False onde true conrresponde aos países encontrados na base 2, que são os países que "Assinaram" o Acordo de Paris.

5.1 Inclusão de coluna na base de dados 1. Emission_Country_Year referente ao Acordo de Paris.

#Criação da coluna 'kyoto' que vale True se 
Emission_Country_Year['Paris'] = Emission_Country_Year['country'].isin(Paris_Agreement_Countries['Participant'])
Emission_Country_Year

# Importar pacote de autenticação do Google e biblioteca para conexão entre Pandas e BigQuery
from google.oauth2 import service_account
import pandas_gbq as pd_gbq

# Autenticar acesso ao BigQuery
acesso = service_account.Credentials.from_service_account_info(
{
  "type": "service_account",
  "project_id": "blueshift-academy",
  "private_key_id": "41e1caa0b572364b9a1a2e0c0062da37bbcbe7ed",
  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCuSbntsF3S72YR\noodN6Fk9DiSbOv6K9wmcUjRCd5prKrfkXehu2D1DpjlIQkYCeOFMeNU4FdHrsKEJ\nQy9CL5m8BLlzhZ822NKFpTBaEC6e4zQzBXlzhT5EaDbgHntculSnLpph8v/gHfWf\ndWckGgI2PUH1pOLsCUCbFQA22D8WQMPlVjBx+yCRVkU65x17l7XrlEBYw/ko/duH\n3v/GJQYU3qaX774T6BGBocj2CoqR5tgE93CWHFw5iZXuUm6d49aMkU3VZ8gH+B/M\n5BucMm6La6zp1gOGe6ureHY48OLugfOuAnmZhOcD6eXuIpDf2s4TqCf6r5xT3lls\nDRJ6R0YxAgMBAAECggEAHWrcytl8sKSPRBU8hMIdw+At/hSCT8bcNKosX3asFgAX\nG2DdYU424wL8wq1K/1JTZtKzXHzUWpWMzNEt+hwqNPHQModAhSOqPvwPkfgOwmtD\njYQUSJMaHDZB4ujKNi/OJ4zGapd3KGcuJOzOTKnmKy0CQ9nBooj6QmLUTy2HUnQz\nLEA7KCbEhOVv0V75JeXIK6RUJD/Kcxt4oONDxQcGZ7qKAfZCvBRDFrryByin2JeN\nbfl2YbOCbCzJfZCc8b4PwoWY+6cuHWoHW1QqEq4vei1PfuL690xtUB1Tb4M3ieuO\nZOLIPYs6lsTCV1yXe08fb7ts/HrU0VYPI4KEzfexxQKBgQDcw/bHFD//iE0W2Oz4\nS+yXtHg+Xwkrd1NyRGuohAYOMRbRj3S/lIk60jU03xFI+6Uyma3HNy2zMbZaGVhJ\n7l0w42zftxQwbORtu8IZnJTvpSLx3VcQ5i+tGKpyN8/h7BGdXUDeLtEyMq98EOTs\nmqFJL015yb6MZo2sD9Da2gzB1QKBgQDKGsj2WV1Q0kfvCh1PcTLsnERDjII+X3hT\nHnt9QIRtjy7vBzxwunc4sfsKJn+MTDsPwP/opqv1tz4nzKqc4ZVvcOKrc+M0WwsR\nkDEykrR3kn6jScRcvDuScA13Xoka21wnt/+BYppK3SCJjjWqtNcNZb+2C2iBx60c\nM4GKQ26E7QKBgCUnviLnX5YMRJ0xyUBI1vr5Xrtwl0R/gsXaihwEjFhbHrIKaOXN\nhAMoWdgaKviTkcwgXUUIHmLoXP5xAZD/0yNlQ2+rinRT/2YQkaxp1A85RdZbIEox\n05hYbGdoyU4Gwzgu97ntOzwtl6dT2JORus9sbKewsqdCo1lwGACdbkQ5AoGBALRB\nd4Pg6I2xP19AEixMtjK6nVfx9ihTViosQc/Qv0xILIAcYBfa94Klm1XOSGyFgRfn\n91c1cKHTMailrnYF7RZcBI8O6VJE8XZUuMx2dGdBTZmtrWlYqXlOX3Rb6MT/aQ4z\n2bs4iozC9MY7BDqg2D2mxnpDao513TwN5DfBNM1xAoGBAMpHVOeEcYaPbAMfCJrL\nYRdIKzyBXWWzZoA2iCz4QI2HQ9uiLQQogGxQbTaQaKxQKnbY+1kAfEkfhgJ+K4WB\nxvUpyLhGZ4n05tr1ySKt5r1dMhjKdWI+PzUO+VnC+MvqV6hStu98uH3nLABC7q1n\n/WtWzmm5JzEcij1X/9/mw+cK\n-----END PRIVATE KEY-----\n",
  "client_email": "725609614257-compute@developer.gserviceaccount.com",
  "client_id": "116558903634579342107",
  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
  "token_uri": "https://oauth2.googleapis.com/token",
  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/725609614257-compute%40developer.gserviceaccount.com"
})

# Atualizar o cache de credenciais na memória
pd_gbq.context.credentials = acesso
pd_gbq.context.project = "blueshift-academy"

# Nome do projeto onde o Conjunto de Dados está hospedado
Projeto= 'blueshift-academy'

# Inserir o DataFrame no BigQuery - (NomeDataFrame, ConjuntoDados.NomeTabela, ...)
pd_gbq.to_gbq(Emission_Country_Year, 'Paris.Emission_Country_Year', project_id=Projeto, if_exists='replace')
pd_gbq.to_gbq(Paris_Agreement_Countries, 'Paris.Paris_Agreement_Countries', project_id=Projeto, if_exists='replace')
pd_gbq.to_gbq(Emission_Country_Year_in, 'Paris.Emission_Country_Year_in', project_id=Projeto, if_exists='replace')
pd_gbq.to_gbq(Paris_Agreement_Countries_in, 'Paris.Paris_Agreement_Countries_in', project_id=Projeto, if_exists='replace')

# Autor: Greyce da Costa Silva
