==> Scripts/3. Country_Region_Continent(Finalizado).ipynb <==
# Base Países por região

Este notebook tem o intuito de relacionar as bases 1 e 2 conforme descritos abaixo:

1. "Emission_Country_Year" - Base principal do projeto, pois contén os dados quantitativos relacionados a emissão de gases poluentes por País e Ano, esta base registra dados de 1850 a 2019 de 236 países.

2. "Country_Continent_Region" - Base auxiliar do projeto, pois contém informações de Paíse e continentes o quais pertencem, esta base foi preparada para conter 236 países assim como a base principal do projeto.

As relações das bases se darão pelas colunas Base 1, Coluna "country" e Base 2, coluna "country" que são comuns em ambas as bases, Foram necessárias algumas trtativas na Base 2 a fim de deixa-la o mais proximo possível da Base 1.


# Instalar os programas e bibliotecas que serão utilizados
!sudo apt install
!pip install kaggle
!sudo apt-get install unzip
!pip install pandas
import pandas as pd

1. Base de Dados 1 - "Emission_Country_Year"

#Leitura de arquivo salvo na VM
Emission_Country_Year = pd.read_csv('/home/jupyter/Emission_Country_Year.csv')
#exibição do dataframe
Emission_Country_Year.columns

#Criação de dataframe com 2 colunas de Emission_Country_Year
Country_IsoCode = Emission_Country_Year[['iso_code', 'country']]
#Drop todos os valores nulos da coluna 'country' 
Country_IsoCode.dropna(subset=['country'], inplace=True)

#Verificando se existe valor nulo em iso_code
Country_IsoCode[Country_IsoCode['iso_code'].isna()].count()

2. Base de Dados 2 - "Country_Continent_Region"

# Importar a biblioteca do Kaggle para acessar a API
import kaggle

# Alterar as permissões de acesso ao arquivo e executá-lo
api_token = {"username":"mayaragomesl","key":"44c49a9ebbe99a1a94eb26d16313f588"}

import json
    
with open('/home/jupyter/.kaggle/kaggle.json', 'w') as file:
    json.dump(api_token, file)

!chmod 600 ~/.kaggle/kaggle.json

#Download Dataset
!kaggle datasets download -d statchaitya/countrycontinent --force

#Descompactando o arquivo countrycontinent.zip no diretorio informado
!unzip -d /home/jupyter/countrycontinent.zip

#Criação de dataframe a partir do arquivo csv informado
country_continent = pd.read_csv('/home/jupyter/countryContinent.csv')
#criação de novo dataframe com os dados anteriores ao tratamento
Country_Continent_Region_in = country_continent
#exibição do dataframe
country_continent

#Criação de novo dataframe a partir de country_continent
Country_Continent_Region = country_continent[['country', 'code_3','country_code','continent','sub_region']]
#Substituição de espaço por nada ao final de cada país na coluna 'country'
Country_Continent_Region.country = Country_Continent_Region.country.str.replace(' $', '')
#Drop todos os valores na na coluna 'country'
Country_Continent_Region.dropna(subset = ['country'], inplace=True)
#exibição do dataframe
Country_Continent_Region.head()

3. Alteração na Base de Dados 2 - "Country_Continent_Region" conforme em Base 1.

#Substituição dos nomes na base Country_Continent_Region conforme o dataframe Emission_Country_Year
Country_Continent_Region.replace(to_replace={'country': 'Bolivia (Plurinational State of)'}, value='Bolivia', inplace=True)
Country_Continent_Region.replace(to_replace={'country': 'Bonaire, Sint Eustatius and Saba'}, value='Bonaire Sint Eustatius and Saba', inplace=True)
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Brunei Darussalam", "Brunei")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Cabo Verde", "Cape Verde")
Country_Continent_Region.replace(to_replace={'country': "Congo (Democratic Republic of the)"}, value="Democratic Republic of Congo", inplace=True)
Country_Continent_Region.replace(to_replace={'country': "C�te d'Ivoire"}, value="Cote d'Ivoire", inplace=True)
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Czech Republic", "Czechia")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Faroe Islands", "Faeroe Islands")
Country_Continent_Region.replace(to_replace={'country': "Iran (Islamic Republic of)"}, value="Iran", inplace=True)
Country_Continent_Region.replace(to_replace={'country': "Korea (Democratic People's Republic of)"}, value="North Korea", inplace=True)
Country_Continent_Region.replace(to_replace={'country': "Korea (Republic of)"}, value="South Korea", inplace=True)
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Lao People's Democratic Republic", "Laos")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Macedonia (the former Yugoslav Republic of)", "North Macedonia")
Country_Continent_Region.replace(to_replace={'country': "Moldova (Republic of)"}, value="Moldova", inplace=True)
Country_Continent_Region.replace(to_replace={'country': "Micronesia (Federated States of)"}, value="Micronesia", inplace=True)
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Russian Federation", "Russia")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Swaziland", "Eswatini")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Syrian Arab Republic", "Syria")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Timor-Leste", "Timor")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("United Kingdom of Great Britain and Northern Ireland", "United Kingdom")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("United States of America", "United States")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Viet Nam", "Vietnam")
Country_Continent_Region.replace(to_replace={'country': "Wallis and Futuna"}, value="Wallis and Futuna Islands", inplace=True)
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Palestine, State of", "Palestine")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Tanzania, United Republic of", "Tanzania")
Country_Continent_Region.country = Country_Continent_Region.country.str.replace("Taiwan, Province of China", "Taiwan")
Country_Continent_Region.replace(to_replace={'country': "Venezuela (Bolivarian Republic of)"}, value="Venezuela", inplace=True)
Country_Continent_Region.replace(to_replace={'country': "Saint Helena, Ascension and Tristan da Cunha"}, value="Saint Helena", inplace=True)
Country_Continent_Region.replace(to_replace={'country': "Virgin Islands (British)"}, value="British Virgin Islands", inplace=True)
Country_Continent_Region.replace(to_replace={'country': "Macedonia (the former Yugoslav Republic of)"}, value="North Macedonia", inplace=True)

#Criação do dataframa com todas as linhas existentes no dataframe 1 e 2
Isocode_Region = Country_Continent_Region[Country_Continent_Region['country'].isin(Country_IsoCode['country'])]
#exibição do dataframe
Isocode_Region.head()

#Criação de nova coluna no dataframe 2 com valores True caso o nome do país esteja na base 1
Country_Continent_Region['IsInCountry_IsoCode']= Country_Continent_Region['country'].isin(Country_IsoCode['country'])
#Seleção de valores False na coluna 'IsInCountry_IsoCode', ou seja não está na base 1
indexNames = Country_Continent_Region[(Country_Continent_Region['IsInCountry_IsoCode'] == False)].index 
#Drop todos os valores indexados
Country_Continent_Region.drop(indexNames, inplace=True)
#Drop coluna 'IsInCountry_IsoCode'
Country_Continent_Region.drop(['IsInCountry_IsoCode'],axis=1, inplace=True)
#exibição do dataframe
Country_Continent_Region.head()

4. Inlcusão de dados na Base de Dados 2 - "Country_Continent_Region" conforme em Base 1.

#Declaração de chaves e valores a serem incluídos no dataframe que será criado
d = {'country': ['Africa', 'Asia', 'Asia (excl. China & India)', 'EU-27', 'EU-28',
       'Europe', 'Europe (excl. EU-27)', 'Europe (excl. EU-28)',
       'French Equatorial Africa', 'French West Africa',
       'International transport', 'Kosovo', 'Kuwaiti Oil Fires',
       'Leeward Islands', 'North America', 'North America (excl. USA)',
       'Oceania', 'Panama Canal Zone', 'Ryukyu Islands', 'South America',
       'St. Kitts-Nevis-Anguilla', 'World'],
     'code_3': ['NaN','NaN' ,'NaN','NaN' ,'NaN','NaN','NaN' ,'NaN','NaN' ,'NaN','NaN' ,'NaN','NaN' ,'NaN','NaN','NaN' ,'NaN','NaN' ,'NaN','NaN','NaN' ,'NaN'], 
     'country_code': ['NaN','NaN' ,'NaN','NaN' ,'NaN','NaN','NaN' ,'NaN','NaN' ,'NaN','NaN' ,'NaN','NaN' ,'NaN','NaN','NaN' ,'NaN','NaN' ,'NaN','NaN','NaN' ,'NaN'], 
     'continent': ['Africa', 'Asia', 'Asia', 'Europe', 'Europe',
       'Europe', 'Europe', 'Europe',
       'Africa', 'Africa',
       'International transport', 'Europe', 'Asia',
       'America', 'America', 'America',
       'Oceania', 'America', 'Asia', 'America',
       'America', 'World'], 
     'sub_region': ['Africa', 'Asia', 'Asia (excl. China & India)', 'EU-27', 'EU-28',
       'Europe', 'Europe (excl. EU-27)', 'Europe (excl. EU-28)',
       'French Equatorial Africa', 'French West Africa',
       'International transport', 'Kosovo', 'Kuwaiti Oil Fires',
       'Caribbean', 'North America', 'North America (excl. USA)',
       'Oceania', 'North America', 'Asia', 'South America',
       'Caribbean', 'World']}
#Criação do dataframa df com os valores declarados em d
df = pd.DataFrame(data=d)
#exibição do dataframe
df.head()

#Concatenação dos dataframes Country_Continent_Region e df em Country_Continent_Region
Country_Continent_Region = pd.concat([Country_Continent_Region,df])
#exibição do dataframe
Country_Continent_Region

#Verificaçao de continent com valor na
Country_Continent_Region[Country_Continent_Region['continent'].isna()]

#Alteração de valor na para 'Asia'
Country_Continent_Region['continent'].fillna('Asia', inplace=True)
#Alteração de valor na para 'Americas' para America
Country_Continent_Region.replace(to_replace={'continent': 'Americas'}, value='America', inplace=True)
#Verificação de continent com valor na
Country_Continent_Region[Country_Continent_Region['continent'].isna()]

#Verificação de continente
Country_Continent_Region['continent'].unique()

5. Comparação das base 1 - "Emission_Country_Year" e 2 - "Country_Continent_Region".

#Verificação dos paises em que estão nos 2 dataframes
Isocode_Region = Country_Continent_Region[Country_Continent_Region['country'].isin(Country_IsoCode['country'])]
#exibição do dataframe- 
Isocode_Region

#Verificação dos paises em que estão em Country_Continent_Region porém não estão no dataframe Country_IsoCode
NaoIsocode_Region = Country_Continent_Region[Country_Continent_Region['country'].isin(Country_IsoCode['country']) == False]
#contagem de países do dataframe 'NaoIsocode_Region'
NaoIsocode_Region.count()

#Paises presentes em Country_IsoCode w não presentes em Country_Continent_Region
Isocode_NaoRegion = Country_IsoCode[Country_IsoCode['country'].isin(Country_Continent_Region['country']) == False]
#Verificação destes valores
Isocode_NaoRegion.country.unique()

6. Conexão com o BigQuery e armazenamento dos DataFrames

# Importar pacote de autenticação do Google e biblioteca para conexão entre Pandas e BigQuery
from google.oauth2 import service_account
import pandas_gbq as pd_gbq

# Autenticar acesso ao BigQuery
acesso = service_account.Credentials.from_service_account_info(
{
  "type": "service_account",
  "project_id": "blueshift-academy",
  "private_key_id": "41e1caa0b572364b9a1a2e0c0062da37bbcbe7ed",
  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCuSbntsF3S72YR\noodN6Fk9DiSbOv6K9wmcUjRCd5prKrfkXehu2D1DpjlIQkYCeOFMeNU4FdHrsKEJ\nQy9CL5m8BLlzhZ822NKFpTBaEC6e4zQzBXlzhT5EaDbgHntculSnLpph8v/gHfWf\ndWckGgI2PUH1pOLsCUCbFQA22D8WQMPlVjBx+yCRVkU65x17l7XrlEBYw/ko/duH\n3v/GJQYU3qaX774T6BGBocj2CoqR5tgE93CWHFw5iZXuUm6d49aMkU3VZ8gH+B/M\n5BucMm6La6zp1gOGe6ureHY48OLugfOuAnmZhOcD6eXuIpDf2s4TqCf6r5xT3lls\nDRJ6R0YxAgMBAAECggEAHWrcytl8sKSPRBU8hMIdw+At/hSCT8bcNKosX3asFgAX\nG2DdYU424wL8wq1K/1JTZtKzXHzUWpWMzNEt+hwqNPHQModAhSOqPvwPkfgOwmtD\njYQUSJMaHDZB4ujKNi/OJ4zGapd3KGcuJOzOTKnmKy0CQ9nBooj6QmLUTy2HUnQz\nLEA7KCbEhOVv0V75JeXIK6RUJD/Kcxt4oONDxQcGZ7qKAfZCvBRDFrryByin2JeN\nbfl2YbOCbCzJfZCc8b4PwoWY+6cuHWoHW1QqEq4vei1PfuL690xtUB1Tb4M3ieuO\nZOLIPYs6lsTCV1yXe08fb7ts/HrU0VYPI4KEzfexxQKBgQDcw/bHFD//iE0W2Oz4\nS+yXtHg+Xwkrd1NyRGuohAYOMRbRj3S/lIk60jU03xFI+6Uyma3HNy2zMbZaGVhJ\n7l0w42zftxQwbORtu8IZnJTvpSLx3VcQ5i+tGKpyN8/h7BGdXUDeLtEyMq98EOTs\nmqFJL015yb6MZo2sD9Da2gzB1QKBgQDKGsj2WV1Q0kfvCh1PcTLsnERDjII+X3hT\nHnt9QIRtjy7vBzxwunc4sfsKJn+MTDsPwP/opqv1tz4nzKqc4ZVvcOKrc+M0WwsR\nkDEykrR3kn6jScRcvDuScA13Xoka21wnt/+BYppK3SCJjjWqtNcNZb+2C2iBx60c\nM4GKQ26E7QKBgCUnviLnX5YMRJ0xyUBI1vr5Xrtwl0R/gsXaihwEjFhbHrIKaOXN\nhAMoWdgaKviTkcwgXUUIHmLoXP5xAZD/0yNlQ2+rinRT/2YQkaxp1A85RdZbIEox\n05hYbGdoyU4Gwzgu97ntOzwtl6dT2JORus9sbKewsqdCo1lwGACdbkQ5AoGBALRB\nd4Pg6I2xP19AEixMtjK6nVfx9ihTViosQc/Qv0xILIAcYBfa94Klm1XOSGyFgRfn\n91c1cKHTMailrnYF7RZcBI8O6VJE8XZUuMx2dGdBTZmtrWlYqXlOX3Rb6MT/aQ4z\n2bs4iozC9MY7BDqg2D2mxnpDao513TwN5DfBNM1xAoGBAMpHVOeEcYaPbAMfCJrL\nYRdIKzyBXWWzZoA2iCz4QI2HQ9uiLQQogGxQbTaQaKxQKnbY+1kAfEkfhgJ+K4WB\nxvUpyLhGZ4n05tr1ySKt5r1dMhjKdWI+PzUO+VnC+MvqV6hStu98uH3nLABC7q1n\n/WtWzmm5JzEcij1X/9/mw+cK\n-----END PRIVATE KEY-----\n",
  "client_email": "725609614257-compute@developer.gserviceaccount.com",
  "client_id": "116558903634579342107",
  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
  "token_uri": "https://oauth2.googleapis.com/token",
  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/725609614257-compute%40developer.gserviceaccount.com"
})

# Atualizar o cache de credenciais na memória
pd_gbq.context.credentials = acesso
pd_gbq.context.project = "blueshift-academy"

# Nome do projeto onde o Conjunto de Dados está hospedado
Projeto = 'blueshift-academy'

# Inserir o DataFrame no BigQuery - (NomeDataFrame, ConjuntoDados.NomeTabela, ...)
pd_gbq.to_gbq(Country_Continent_Region, 'Continent.Country_Continent_Region', project_id=Projeto, if_exists='replace')
pd_gbq.to_gbq(Country_Continent_Region_in, 'Continent.Country_Continent_Region_in', project_id=Projeto, if_exists='replace')

# Autor: Matheus Gois
